+++ sdk/aot/azure-aot-graalvm-support/src/main/java/com/azure/aot/graalvm/support/implementation/features/AppConfigFeature.java
@@ -21,7 +21,7 @@ public class AppConfigFeature implements GraalVMFeature {
     @Override
     public Set<String[]> getDynamicProxies() {
         return setsOf(
-                interfaces("com.azure.data.appconfiguration.ConfigurationService")
+                interfaces("com.azure.data.appconfiguration.implementation.ConfigurationService")
         );
     }
 }