+++ sdk/cosmos/azure-cosmos/src/test/java/com/azure/cosmos/CosmosDiagnosticsTest.java
@@ -241,6 +241,9 @@ public class CosmosDiagnosticsTest extends TestSuiteBase {
             assertThat(diagnostics).contains("\"statusCode\":404");
             assertThat(diagnostics).contains("\"operationType\":\"Read\"");
             assertThat(diagnostics).contains("\"userAgent\":\"" + Utils.getUserAgent() + "\"");
+            assertThat(diagnostics).contains("\"exceptionMessage\":\"Entity with the specified id does not exist in the system.");
+            assertThat(diagnostics).contains("\"exceptionResponseHeaders\"");
+            assertThat(diagnostics).doesNotContain("\"exceptionResponseHeaders\": \"{}\"");
             assertThat(diagnostics).containsAnyOf(
                 "\"machineId\":\"" + tempMachineId + "\"", // logged machineId should be static uuid or
                 "\"machineId\":\"" + ClientTelemetry.getMachineId(null) + "\"" // the vmId from Azure
@@ -316,6 +319,9 @@ public class CosmosDiagnosticsTest extends TestSuiteBase {
             } catch (CosmosException e) {
                 diagnostics = e.getDiagnostics().toString();
                 assertThat(diagnostics).contains("\"backendLatencyInMs\"");
+                assertThat(diagnostics).contains("\"exceptionMessage\":\"[\\\"Resource with specified id or name already exists.\\\"]\"");
+                assertThat(diagnostics).contains("\"exceptionResponseHeaders\"");
+                assertThat(diagnostics).doesNotContain("\"exceptionResponseHeaders\": \"{}\"");
                 validateTransportRequestTimelineDirect(e.getDiagnostics().toString());
             }
         } finally {
@@ -670,6 +676,9 @@ public class CosmosDiagnosticsTest extends TestSuiteBase {
             assertThat(exception.getDiagnostics().getDuration()).isNotNull();
             assertThat(diagnostics).contains("\"backendLatencyInMs\"");
             assertThat(diagnostics).containsPattern("(?s).*?\"activityId\":\"[^\\s\"]+\".*");
+            assertThat(diagnostics).contains("\"exceptionMessage\":\"[\\\"Resource Not Found.");
+            assertThat(diagnostics).contains("\"exceptionResponseHeaders\"");
+            assertThat(diagnostics).doesNotContain("\"exceptionResponseHeaders\":null");
             isValidJSON(diagnostics);
             validateTransportRequestTimelineDirect(diagnostics);
             validateRegionContacted(createResponse.getDiagnostics(), client.asyncClient());
@@ -705,6 +714,7 @@ public class CosmosDiagnosticsTest extends TestSuiteBase {
             String diagnostics = exception.getDiagnostics().toString();
             assertThat(exception.getStatusCode()).isEqualTo(HttpConstants.StatusCodes.BADREQUEST);
             assertThat(diagnostics).contains("\"connectionMode\":\"DIRECT\"");
+            assertThat(diagnostics).contains("\"exceptionMessage\":\"TestBadRequest\"");
             assertThat(diagnostics).doesNotContain(("\"resourceAddress\":null"));
             assertThat(diagnostics).contains("\"resourceType\":\"DocumentCollection\"");
             assertThat(exception.getDiagnostics().getContactedRegionNames()).isNotEmpty();
@@ -818,6 +828,12 @@ public class CosmosDiagnosticsTest extends TestSuiteBase {
             fail("expected to fail due to 409");
         } catch (CosmosException e) {
             // no request payload and no response payload
+            logger.info("Diagnostics are : {}", e.getDiagnostics());
+            String diagnostics = e.getDiagnostics().toString();
+            assertThat(diagnostics).contains("\"exceptionMessage\":\"[\\\"Resource with specified id or name already exists.\\\"]\"");
+            assertThat(diagnostics).contains("\"exceptionResponseHeaders\"");
+            assertThat(diagnostics).doesNotContain("\"exceptionResponseHeaders\": \"{}\"");
+
             validate(e.getDiagnostics(), testItemLength, 0);
         }
 
@@ -840,9 +856,7 @@ public class CosmosDiagnosticsTest extends TestSuiteBase {
         try {
 
             DirectConnectionConfig connectionConfig = DirectConnectionConfig.getDefaultConfig();
-            if (connectionStateListenerEnabled) {
-                connectionConfig.setConnectionEndpointRediscoveryEnabled(true);
-            }
+            connectionConfig.setConnectionEndpointRediscoveryEnabled(connectionStateListenerEnabled);
 
             client1 = new CosmosClientBuilder()
                 .endpoint(TestConfigurations.HOST)
@@ -915,8 +929,7 @@ public class CosmosDiagnosticsTest extends TestSuiteBase {
         JsonNode storeResult = responseStatisticsList.get(0).get("storeResult");
         assertThat(storeResult).isNotNull();
 
-        boolean hasPayload = storeResult.get("exception").isNull();
-        assertThat(storeResult.get("channelTaskQueueSize").asInt(-1)).isGreaterThan(0);
+        assertThat(storeResult.get("channelTaskQueueSize").asInt(-1)).isGreaterThanOrEqualTo(0);
         assertThat(storeResult.get("pendingRequestsCount").asInt(-1)).isGreaterThanOrEqualTo(0);
 
         JsonNode serviceEndpointStatistics = storeResult.get("serviceEndpointStatistics");
@@ -968,7 +981,7 @@ public class CosmosDiagnosticsTest extends TestSuiteBase {
         assertThat(responseStatisticsList.size()).isGreaterThan(0);
         JsonNode storeResult = responseStatisticsList.get(0).get("storeResult");
 
-        boolean hasPayload = storeResult.get("exception").isNull();
+        boolean hasPayload = storeResult.get("exceptionMessage") == null;
         assertThat(storeResult).isNotNull();
         assertThat(storeResult.get("rntbdRequestLengthInBytes").asInt(-1)).isGreaterThan(expectedRequestPayloadSize);
         assertThat(storeResult.get("rntbdRequestLengthInBytes").asInt(-1)).isGreaterThan(expectedRequestPayloadSize);
@@ -1005,9 +1018,7 @@ public class CosmosDiagnosticsTest extends TestSuiteBase {
             assertThat(writeResourceResponse.getDiagnostics().toString()).contains("addressResolutionStatistics");
             assertThat(writeResourceResponse.getDiagnostics().toString()).contains("\"inflightRequest\":false");
             assertThat(writeResourceResponse.getDiagnostics().toString()).doesNotContain("endTime=\"null\"");
-            assertThat(writeResourceResponse.getDiagnostics().toString()).contains("\"errorMessage\":null");
-            assertThat(writeResourceResponse.getDiagnostics().toString()).doesNotContain("\"errorMessage\":\"io.netty" +
-                ".channel.AbstractChannel$AnnotatedConnectException: Connection refused: no further information");
+            assertThat(writeResourceResponse.getDiagnostics().toString()).contains("\"exceptionMessage\":null");
 
             client2 = new CosmosClientBuilder()
                 .endpoint(TestConfigurations.HOST)
@@ -1048,8 +1059,7 @@ public class CosmosDiagnosticsTest extends TestSuiteBase {
             assertThat(readResourceResponse.getDiagnostics().toString()).contains("addressResolutionStatistics");
             assertThat(readResourceResponse.getDiagnostics().toString()).contains("\"inflightRequest\":false");
             assertThat(readResourceResponse.getDiagnostics().toString()).doesNotContain("endTime=\"null\"");
-            assertThat(readResourceResponse.getDiagnostics().toString()).contains("\"errorMessage\":null");
-            assertThat(readResourceResponse.getDiagnostics().toString()).contains("\"errorMessage\":\"io.netty" +
+            assertThat(readResourceResponse.getDiagnostics().toString()).contains("\"exceptionMessage\":\"io.netty" +
                 ".channel.AbstractChannel$AnnotatedConnectException: Connection refused");
         } catch (Exception ex) {
             logger.error("Error in test addressResolutionStatistics", ex);
@@ -1098,7 +1108,6 @@ public class CosmosDiagnosticsTest extends TestSuiteBase {
     }
 
     private void validateTransportRequestTimelineGateway(String diagnostics) {
-        assertThat(diagnostics).contains("\"eventName\":\"connectionConfigured\"");
         assertThat(diagnostics).contains("\"eventName\":\"connectionConfigured\"");
         assertThat(diagnostics).contains("\"eventName\":\"requestSent\"");
         assertThat(diagnostics).contains("\"eventName\":\"transitTime\"");