+++ sdk/textanalytics/azure-ai-textanalytics/src/main/java/com/azure/ai/textanalytics/implementation/models/TasksStateTasksDetails.java
@@ -8,4 +8,6 @@ import com.azure.core.annotation.Immutable;
 
 /** The TasksStateTasksDetails model. */
 @Immutable
-public final class TasksStateTasksDetails extends TaskState {}
+public final class TasksStateTasksDetails extends TaskState {
+
+}